Elscreen: Removed and replaced by using make-frame in wl etc.
elscreen does not work well with some important packages such as multi-eshell,
is not being actively maintained and has to be patched by hand to work with
emacs-23.2 so it is becoming a liability.  On stumpwm the handling of frames is
so effective and emacs-like that there is no real need for elscreen anymore.